Hospital management assignment.
Yet to do:
Read initial file and enqueue items.
Currently, heap is having the entire node info. have to make it to have only pid of the patient.
This means, patientRecord.java should implement an id-based lookup wich will return age.
This function should be used in the consultQueue heap.